% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realscrape.R
\name{scrape_sbc}
\alias{scrape_sbc}
\title{scrape_sbc}
\usage{
scrape_sbc(sbc, B = 1, P = 12, noisy = TRUE)
}
\arguments{
\item{sbc}{A list of suburbs by council
in the same format as the example list
\code{STA3LM_suburbs_by_council}}

\item{B}{Number of boards (i.e., pages of ads) to scape per suburb.}

\item{P}{Number of ad pages per board (determined by inspecting the
AU House Prices website).}

\item{noisy}{Logical. If \code{TRUE} then the function prints
a message each time a suburb is completed.}
}
\value{
A tibble of house sales data.
}
\description{
Scrape house sales data from a list of suburbs by council.
}
\details{
This function relies on a loop which repeatedly
calls \code{\link[realscrape]{scrape_suburb}}.
}
\examples{
View( STA3LM_suburbs_by_council )

# Normally we would set P = 12 (12 ad pages per board)
data <- scrape_sbc( STA3LM_suburbs_by_council, B = 1, P = 1 )
data_clean <- tidyr::drop_na( data )

}
