% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realscrape.R
\name{scrape_suburb}
\alias{scrape_suburb}
\title{scrape_suburb}
\usage{
scrape_suburb(suburb, postcode, B = 1, P = 12)
}
\arguments{
\item{suburb}{A character string giving the suburb name
with empty spaces replaced with '+', and first letter of each
word capitalised.}

\item{postcode}{A character string giving the 4 digit suburb
postcode.}

\item{B}{Number of boards (i.e., pages of ads) to scape per suburb.}

\item{P}{Number of ad pages per board (determined by inspecting the
AU House Prices website).}
}
\value{
A tibble containing house sales data with missing values
represented by \code{NA}.
}
\description{
Scrape house sales data from a suburb.
}
\examples{

# Normally we would set P = 12 (12 ad pages per board)
data <- scrape_suburb( suburb = "Bayswater+North", postcode = "3153", P = 1 )


}
