% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realscrape.R
\name{scrape_suburb}
\alias{scrape_suburb}
\title{scrape_suburb}
\usage{
scrape_suburb(suburb, B = 1, P = 1)
}
\arguments{
\item{suburb}{A vector whose first element is the suburb name
with empty spaces replaced with '+' and first letter of each
word capitalised. The second element is the suburb postcode.}

\item{B}{Number of boards (i.e., pages of ads) to scape per suburb.}

\item{P}{Number of ad pages per board (determined by inspecting the
AU House Prices website).}
}
\value{
A tibble containing house sales data with missing values
represented by \code{NA}.
}
\description{
scrape_suburb
}
\examples{

suburbs_by_council <- list(
  MAROONDAH = list(
    c("Bayswater+North",      "3153"),
    c("Croydon"        ,      "3136")),
   KNOX = list(
    c("Bayswater"           , "3153"),
    c("Boronia"             , "3155")))

data <- data.frame()

for ( council in names(suburbs_by_council)  ) {
  for ( suburb in suburbs_by_council[[council]] ) {

    data <- suburb \%>\%
      scrape_suburb() \%>\%
      rbind(data)
  }
}


}
